Extract logic to useAdminSidebarWidth hook
