+++ packages/js/admin-layout/changelog/fix-calypso-adminmenu
@@ -1,0 +1,4 @@
+Significance: minor
+Type: add
+
+Add useAdminSidebarWidth hook
