+++ packages/js/admin-layout/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-admin-sidebar-width';
