+++ packages/js/admin-layout/src/index.ts
@@ -1,2 +1,3 @@
 export * from './plugins';
 export * from './components';
+export * from './hooks';
